major refactor (core `tensor.py`/`dataset.py`)
