+++ hub/core/chunk_engine/read.py
@@ -9,71 +9,6 @@ from hub.util.array import normalize_and_batchify_shape
 from hub.util.keys import get_index_map_key, get_tensor_meta_key
 
 
-def read_tensor_meta(key: str, storage: StorageProvider) -> dict:
-    return pickle.loads(storage[get_tensor_meta_key(key)])
-
-
-def read_index_map(key: str, storage: StorageProvider) -> List[dict]:
-    return pickle.loads(storage[get_index_map_key(key)])
-
-
-def read_dataset_meta(storage: StorageProvider) -> dict:
-    return pickle.loads(storage[constants.META_FILENAME])
-
-
-def tensor_meta_from_array(
-    array: np.ndarray, batched: bool, chunk_size: int = constants.DEFAULT_CHUNK_SIZE
-):
-    array = normalize_and_batchify_shape(array, batched=batched)
-
-    tensor_meta = {
-        "chunk_size": chunk_size,
-        "dtype": array.dtype.name,
-        "length": 0,
-        "min_shape": tuple(array.shape[1:]),
-        "max_shape": tuple(array.shape[1:]),
-        # TODO: add entry in meta for which tobytes function is used and handle mismatch versions for this
-    }
-
-    return tensor_meta
-
-
-def tensor_exists(key: str, storage: StorageProvider) -> bool:
-    """A tensor exists if at the specified `key` and `storage` there is both a meta file and index map."""
-
-    meta_key = get_tensor_meta_key(key)
-    index_map_key = get_index_map_key(key)
-    return meta_key in storage and index_map_key in storage
-
-
-def read_samples_from_tensor(
-    key: str,
-    storage: StorageProvider,
-    array_slice: slice = slice(None),
-) -> np.ndarray:
-    """Read (and unpack) samples from a tensor as an np.ndarray.
-
-    Args:
-        key (str): Key for where the chunks, index_map, and meta are located in `storage` relative to it's root.
-        array_slice (slice): Slice that represents which samples to read. Default = slice representing all samples.
-        storage (StorageProvider): StorageProvider for reading the chunks, index_map, and meta.
-
-    Returns:
-        np.ndarray: Array containing the sample(s) in the `array_slice` slice.
-    """
-
-    meta = read_tensor_meta(key, storage)
-    index_map = read_index_map(key, storage)
-
-    # TODO: read samples in parallel
-    samples = []
-    for index_entry in index_map[array_slice]:
-        array = sample_from_index_entry(key, storage, index_entry, meta["dtype"])
-        samples.append(array)
-
-    return np.array(samples)
-
-
 def sample_from_index_entry(
     key: str, storage: StorageProvider, index_entry: dict, dtype: str
 ) -> np.ndarray:
