+++ hub/core/chunk_engine/tests/test_write_failures.py
@@ -1,7 +1,9 @@
 import numpy as np
 import pytest
-from hub.core.chunk_engine.read import tensor_meta_from_array
-from hub.core.chunk_engine.write import add_samples_to_tensor, create_tensor
+
+from hub.core.meta.tensor_meta import tensor_meta_from_array
+from hub.core.tensor import add_samples_to_tensor, create_tensor
+
 from hub.tests.common import TENSOR_KEY
 from hub.util.exceptions import TensorMetaMismatchError
