+++ hub/core/chunk_engine/write.py
@@ -1,97 +1,11 @@
-import pickle  # TODO: NEVER USE PICKLE
-from typing import Any, Callable, List, Tuple
+from typing import List, Tuple
 from uuid import uuid1
 
 import numpy as np
-from hub.constants import META_FILENAME
 from hub.core.typing import StorageProvider
-from hub.util.array import normalize_and_batchify_shape
-from hub.util.exceptions import TensorMetaMismatchError
-from hub.util.keys import get_chunk_key, get_index_map_key, get_tensor_meta_key
+from hub.util.keys import get_chunk_key
 
 from .chunker import generate_chunks
-from .flatten import row_wise_to_bytes
-from .read import read_index_map, read_tensor_meta, tensor_exists
-
-
-def write_tensor_meta(key: str, storage: StorageProvider, meta: dict):
-    storage[get_tensor_meta_key(key)] = pickle.dumps(meta)
-
-
-def write_index_map(key: str, storage: StorageProvider, index_map: list):
-    index_map_key = get_index_map_key(key)
-    storage[index_map_key] = pickle.dumps(index_map)
-
-
-def write_dataset_meta(storage: StorageProvider, meta: dict):
-    storage[META_FILENAME] = pickle.dumps(meta)
-
-
-def create_tensor(key: str, storage: StorageProvider, meta: dict):
-    """
-    Args:
-        chunk_size (int): Desired length of each chunk.
-    """
-
-    # TODO: docstring
-    # TODO: validate meta
-
-    write_index_map(key, storage, [])
-    write_tensor_meta(key, storage, meta)
-
-
-def add_samples_to_tensor(
-    array: np.ndarray,
-    key: str,
-    storage: StorageProvider,
-    batched: bool = False,
-):
-    """Create a new tensor (if one doesn't already exist), then chunk and write the given array to storage.
-
-    For more on chunking, see the `generate_chunks` method.
-
-    Args:
-        array (np.ndarray): Array to be chunked/written. Batch axis (`array.shape[0]`) is optional, if `array` does have a
-            batch axis, you should pass the argument `batched=True`.
-        key (str): Key for where the chunks, index_map, and meta will be located in `storage` relative to it's root.
-        storage (StorageProvider): StorageProvider for storing the chunks, index_map, and meta.
-
-        batched (bool): If True, the provied `array`'s first axis (`shape[0]`) will be considered it's batch axis.
-            If False, a new axis will be created with a size of 1 (`array.shape[0] == 1`). default=False
-    """
-
-    array = normalize_and_batchify_shape(array, batched=batched)
-
-    if not tensor_exists(key, storage):
-        raise Exception()  # TODO: exceptions.py
-
-    index_map = read_index_map(key, storage)
-    tensor_meta = read_tensor_meta(key, storage)
-    _check_array_and_tensor_are_compatible(tensor_meta, array)
-
-    # TODO: get the tobytes function from meta
-    tobytes = row_wise_to_bytes
-
-    array_length = array.shape[0]
-    for i in range(array_length):
-        sample = array[i]
-
-        # TODO: we may want to call `tobytes` on `array` and call memoryview on that. this may depend on the access patterns we
-        # choose to optimize for.
-        b = memoryview(tobytes(sample))
-
-        index_map_entry = write_bytes(
-            b, key, tensor_meta["chunk_size"], storage, index_map
-        )
-
-        # shape per sample for dynamic tensors (TODO: if strictly fixed-size, store this in meta)
-        index_map_entry["shape"] = sample.shape
-        index_map.append(index_map_entry)
-
-    tensor_meta["length"] += array_length
-
-    write_tensor_meta(key, storage, tensor_meta)
-    write_index_map(key, storage, index_map)
 
 
 def write_bytes(
@@ -198,34 +112,3 @@ def _get_last_chunk(
 def _random_chunk_name() -> str:
     return str(uuid1())
 
-
-def _check_array_and_tensor_are_compatible(tensor_meta: dict, array: np.ndarray):
-    """An array is considered incompatible with a tensor if the `tensor_meta` entries don't match the `array` properties.
-
-    Args:
-        tensor_meta (dict): Tensor meta containing the expected properties of `array`.
-        array (np.ndarray): Candidate array to check compatibility with `tensor_meta`.
-
-    Raises:
-        TensorMetaMismatchError: When `array` properties do not match the `tensor_meta`'s exactly. Also when `len(array.shape)` != len(tensor_meta max/min shapes).
-        NotImplementedError: When `array.shape` does not match for all samples. Dynamic shapes are not yet supported. (TODO)
-    """
-
-    if tensor_meta["dtype"] != array.dtype.name:
-        raise TensorMetaMismatchError("dtype", tensor_meta["dtype"], array.dtype.name)
-
-    sample_shape = array.shape[1:]
-    if len(tensor_meta["min_shape"]) != len(sample_shape):
-        raise TensorMetaMismatchError(
-            "min_shape", tensor_meta["min_shape"], len(sample_shape)
-        )
-    if len(tensor_meta["max_shape"]) != len(sample_shape):
-        raise TensorMetaMismatchError(
-            "max_shape", tensor_meta["max_shape"], len(sample_shape)
-        )
-
-    # TODO: remove these once dynamic shapes are supported
-    if not np.array_equal(tensor_meta["max_shape"], sample_shape):
-        raise NotImplementedError("Dynamic shapes are not supported yet.")
-    if not np.array_equal(tensor_meta["min_shape"], sample_shape):
-        raise NotImplementedError("Dynamic shapes are not supported yet.")
