+++ hub/core/typing.py
@@ -1,1 +1,1 @@
-from hub.core.storage.provider import StorageProvider
+from hub.core.storage import StorageProvider
